# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: 受験マップ API
  version: "1"
servers:
  - url: https://juken-map.com
security:
  - sessionCookie: []
paths:
...
  /api/line/oauth/callback:
    get:
      operationId: completeLineLogin
      summary: LINE Login の戻り先。結果はプロフィール画面へ ?line=<結果> で知らせる
      parameters:
        - { name: state, in: query, required: false, schema: { type: string } }
        - { name: code, in: query, required: false, schema: { type: string } }
        - { name: error, in: query, required: false, schema: { type: string } }
      responses:
        "302":
          description: >-
            プロフィール画面へ（line=connected・cancelled・invalid・expired・friend-required・already-used・failed）。
            セッションが切れていれば、この URL を戻り先にしてログイン画面へ
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-jm_session
      description: ログイン（Go の /api/auth/*、JUK-115）が発行するセッション Cookie。中身の無い 256 ビットの乱数のトークン
    cronBearer:
      type: http
      scheme: bearer
      description: DAILY_NOTIFICATION_SECRET
    simBearer:
      type: http
      scheme: bearer
      description: SIMULATION_SECRET（cron とは別のトークン）
...
  responses:
...
    Forbidden:
      description: 利用を停止されている、デモアカウントの書き込み、など
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: 想定外の失敗。原因はログにだけ残す
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ServerError" }
  schemas:
...
    Error:
      description: ルートが自分で断るときのエラー。error は画面にそのまま出す文言
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ServerError:
      description: 想定外の失敗・混雑・存在しないパス
      type: object
      required: [error, code, reqId]
      properties:
        error: { type: string }
        code:
          type: string
          # FST_ERR_CTP_* は Fastify の code をそのまま使っている（Node の error-handling.ts）
          # FST_ERR_CTP_INVALID_CONTENT_LENGTH は本文が UTF-8 として不正で、読んだ長さが Content-Length と合わないとき（400）
          enum:
            - INTERNAL_ERROR
            - OVERLOADED
            - NOT_FOUND
            - FST_ERR_CTP_BODY_TOO_LARGE
            - FST_ERR_CTP_INVALID_MEDIA_TYPE
            - FST_ERR_CTP_INVALID_CONTENT_LENGTH
            # 利用者単位の回数制限を超えた（429、Go の internal/httpx/user_rate_limit.go）
            - TOO_MANY_REQUESTS
          x-enum-varnames:
            - ServerErrorCodeInternal
            - ServerErrorCodeOverloaded
            - ServerErrorCodeNotFound
            - ServerErrorCodeBodyTooLarge
            - ServerErrorCodeInvalidMediaType
            - ServerErrorCodeInvalidContentLength
            - ServerErrorCodeTooManyRequests
        reqId: { type: string }
...